Gets all tests to run

+++ src/onegov.org/onegov/org/upgrade.py
@@ -3,9 +3,21 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
+from onegov.org.models import Organisation
 
 
-@upgrade_task('Move from town to organisation')
+@upgrade_task('Move from town to organisation', always_run=True)
 def move_town_to_organisation(context):
+
+    # if the organisations table does not exist (other modules), skip
+    if not context.engine.dialect.has_table(context.engine, 'organisations'):
+        return
+
+    # this needs to be an always-run task for now, since on first install the
+    # module is deteced as new and it is assumed that no tasks need to be run
+    # as a result
+    if context.session.query(Organisation).first():
+        return
+
     context.operations.drop_table('organisations')
     context.operations.rename_table('towns', 'organisations')
